edits 14/10

+++ database/migrations/2021_01_13_154951_create_order_items_table.php
@@ -34,7 +34,7 @@ class CreateOrderItemsTable extends Migration
             $table->string('payment_status')->default('pending');
             $table->string('refused_note')->nullable();
 
-            $table->tinyInteger('quantity');
+            $table->smallInteger('quantity');
             $table->double('price');
 
             $table->timestamps();
